letters checker
suppose that you have given a string. It is required to convert each letter to its equivalent uppercase or lowercase according to its position. Such that, if the letter is coming after "," or "." or "!" or "?" it should be uppercase letter or if it was the first letter of paragraph. Otherwise it should be lowercase.